:recycle: Separate squat animation from state
